+++ resources/ReleaseNotes.md
@@ -1,3 +1,5 @@
 # Software
 
 - Make dependence on Parallel Computing Toolbox optional
+- Convert addition of missing row at end of file data into a processing step (`mag.process.Separate`)
+- Add `mag.process.Separate` to HK processing